backend/app/core/security/deps.py
from fastapi import Depends, HTTPException, status, Request, Header, Query
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
from typing import Optional, Tuple, Dict, Any
from datetime import datetime
from app.core.config import settings
from app.db.mongodb import users_collection, devices_collection
from app.core.security.auth import verify_api_key
import logging
from slowapi import Limiter
from slowapi.util import get_remote_address

# Set up limiter for rate limiting
limiter = Limiter(key_func=get_remote_address)

logger = logging.getLogger(__name__)

oauth2_scheme = OAuth2PasswordBearer(
    tokenUrl=f"{settings.API_V1_PREFIX}/auth/login"
)

async def get_api_key_header(
    x_device_key: Optional[str] = Header(None)
) -> Optional[str]:
    """Get API key from header."""
    return x_device_key

async def get_current_user(token: str = Depends(oauth2_scheme)):
    """Verify JWT token and return user data"""
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Invalid authentication credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    
    try:
        payload = jwt.decode(token, settings.JWT_SECRET, algorithms=["HS256"])
        user_id: str = payload.get("sub")
        if user_id is None:
            raise credentials_exception
        
        user = users_collection.find_one({"user_id": user_id})
        if user is None:
            raise credentials_exception
            
        # Remove _id from MongoDB document
        user.pop('_id', None)
        return user
    except JWTError as e:
        logger.warning(f"JWT error: {e}")
        raise credentials_exception
    except Exception as e:
        logger.error(f"Authentication error: {e}")
        raise credentials_exception

async def get_current_active_user(current_user = Depends(get_current_user)):
    """Check if user is active"""
    if not current_user.get("active", True):
        raise HTTPException(status_code=400, detail="Inactive user")
    return current_user

async def get_current_manager(current_user = Depends(get_current_active_user)):
    """Check if user is a manager"""
    if current_user.get("role") != "manager":
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Not enough permissions"
        )
    return current_user

async def get_current_employee(current_user = Depends(get_current_active_user)):
    """Check if user is an employee"""
    if current_user.get("role") != "employee":
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Employee access only"
        )
    return current_user

async def verify_device_api_key(
    request: Request,
    api_key: Optional[str] = Depends(get_api_key_header),
    query_api_key: Optional[str] = Query(None, alias="api_key"),
) -> Dict[str, Any]:
    """
    Verify the device API key
    
    Tries both header and query parameter methods for the API key.
    """
    try:
        # Log debug information
        logger.debug(f"Headers received: {dict(request.headers)}")
        logger.debug(f"API Key from header: {api_key[:4] + '...' if api_key else 'None'}")
        logger.debug(f"API Key from query: {query_api_key[:4] + '...' if query_api_key else 'None'}")
        
        # Use query parameter API key if header is not provided
        if not api_key and query_api_key:
            api_key = query_api_key
            logger.debug(f"Using API key from query parameter")
        
        if not api_key:
            logger.error("No API key provided in headers or query parameters")
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="API key is required",
                headers={"WWW-Authenticate": "ApiKey"},
            )
            
        # Get path parameters for more context in debugging
        path_params = request.path_params
        device_id_from_url = path_params.get("device_id", None)
        logger.debug(f"Device ID from URL path: {device_id_from_url}")
        
        # Get all devices for debugging
        device_list = list(devices_collection.find({}, {"device_id": 1, "api_key": 1, "active": 1}))
        logger.debug(f"Found {len(device_list)} total devices in database")
        for idx, dev in enumerate(device_list):
            logger.debug(f"Device {idx+1}: ID={dev.get('device_id')}, Key={dev.get('api_key')[:6] if dev.get('api_key') else 'None'}..., Active={dev.get('active', False)}")
        
        # First try to find the device with the exact API key
        device = devices_collection.find_one({"api_key": api_key, "active": True})
        
        if not device:
            # Log an error for debugging
            logger.warning(f"Device not found with API key: {api_key[:6]}...")
            
            # If we have a device ID from the URL, try to look up directly (temporary bypass for testing)
            if device_id_from_url:
                device = devices_collection.find_one({"device_id": device_id_from_url})
                if device:
                    logger.warning(f"Using URL device ID as temporary bypass: {device_id_from_url}")
                    return device
            
            # Try to match with the hash instead
            from app.core.security.auth import hash_api_key
            api_key_hash = hash_api_key(api_key)
            device = devices_collection.find_one({"api_key_hash": api_key_hash, "active": True})
            
            if not device:
                logger.error(f"No device found with API key hash: {api_key_hash[:6]}...")
                raise HTTPException(
                    status_code=status.HTTP_401_UNAUTHORIZED,
                    detail="Invalid API key",
                    headers={"WWW-Authenticate": "ApiKey"},
                )
            else:
                logger.debug(f"Device found using API key hash: {device.get('device_id')}")
        else:
            logger.debug(f"Device found using plain API key: {device.get('device_id')}")
        
        # Update last active
        devices_collection.update_one(
            {"_id": device["_id"]},
            {"$set": {"last_active": datetime.utcnow()}}
        )
        
        return device
    except Exception as e:
        logger.error(f"Error in verify_device_api_key: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Authentication error: {str(e)}",
        )
